# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BinoChatTurnRole = typeof BinoChatTurnRole[keyof typeof BinoChatTurnRole];


export const BinoChatTurnRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface BinoChatTurn {
  role: BinoChatTurnRole;
  /** @maxLength 2000 */
  content: string;
}

export type BinoChatInputLanguage = typeof BinoChatInputLanguage[keyof typeof BinoChatInputLanguage];


export const BinoChatInputLanguage = {
  uz: 'uz',
  ru: 'ru',
  en: 'en',
} as const;

export interface BinoChatInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
  language: BinoChatInputLanguage;
  /** @maxItems 16 */
  history?: BinoChatTurn[];
}

export interface BinoChatReply {
  reply: string;
}

export interface BinoLeadInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 7
     * @maxLength 40
     */
  phone: string;
  /**
     * @minLength 3
     * @maxLength 2000
     */
  message: string;
}

export interface BinoLeadReceipt {
  id: number;
  createdAt: string;
}

export interface BinoPhone {
  label: string;
  number: string;
}

export type BinoSiteConfigMetrics = {
  experienceYears: number;
  completedProjects: number;
  satisfiedClients: number;
  warrantyYears: number;
};

export interface BinoSiteConfig {
  phones: BinoPhone[];
  metrics: BinoSiteConfigMetrics;
}