# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: bino
    description: BINO STROY customer interactions
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /bino/chat:
    post:
      operationId: createBinoChatMessage
      tags: [bino]
      summary: Ask the BINO product assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BinoChatInput"
      responses:
        "200":
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BinoChatReply"
        "400":
          description: Invalid chat request
        "503":
          description: Assistant is not configured
  /bino/site-config:
    get:
      operationId: getBinoSiteConfig
      tags: [bino]
      summary: Get editable BINO STROY contact and counter values
      responses:
        "200":
          description: Public site configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BinoSiteConfig"
  /bino/leads:
    post:
      operationId: createBinoLead
      tags: [bino]
      summary: Submit a sales inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BinoLeadInput"
      responses:
        "201":
          description: Inquiry received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BinoLeadReceipt"
        "400":
          description: Invalid inquiry
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    BinoChatTurn:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
          maxLength: 2000
      required: [role, content]
    BinoChatInput:
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 2000
        language:
          type: string
          enum: [uz, ru, en]
        history:
          type: array
          maxItems: 16
          items:
            $ref: "#/components/schemas/BinoChatTurn"
      required: [message, language]
    BinoChatReply:
      type: object
      properties:
        reply:
          type: string
      required: [reply]
    BinoLeadInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        phone:
          type: string
          minLength: 7
          maxLength: 40
        message:
          type: string
          minLength: 3
          maxLength: 2000
      required: [name, phone, message]
    BinoLeadReceipt:
      type: object
      properties:
        id:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, createdAt]
    BinoPhone:
      type: object
      properties:
        label:
          type: string
        number:
          type: string
      required: [label, number]
    BinoSiteConfig:
      type: object
      properties:
        phones:
          type: array
          items:
            $ref: "#/components/schemas/BinoPhone"
        metrics:
          type: object
          properties:
            experienceYears:
              type: integer
            completedProjects:
              type: integer
            satisfiedClients:
              type: integer
            warrantyYears:
              type: integer
          required:
            - experienceYears
            - completedProjects
            - satisfiedClients
            - warrantyYears
      required: [phones, metrics]