# lib/api-zod/src/generated/types/binoChatInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BinoChatInputLanguage } from './binoChatInputLanguage';
import type { BinoChatTurn } from './binoChatTurn';

export interface BinoChatInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  message: string;
  language: BinoChatInputLanguage;
  /** @maxItems 16 */
  history?: BinoChatTurn[];
}