# lib/api-zod/src/generated/types/binoChatTurn.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BinoChatTurnRole } from './binoChatTurnRole';

export interface BinoChatTurn {
  role: BinoChatTurnRole;
  /** @maxLength 2000 */
  content: string;
}