# lib/api-zod/src/generated/types/binoLeadInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BinoLeadInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /**
     * @minLength 7
     * @maxLength 40
     */
  phone: string;
  /**
     * @minLength 3
     * @maxLength 2000
     */
  message: string;
}