# lib/api-zod/src/generated/types/binoSiteConfig.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BinoPhone } from './binoPhone';
import type { BinoSiteConfigMetrics } from './binoSiteConfigMetrics';

export interface BinoSiteConfig {
  phones: BinoPhone[];
  metrics: BinoSiteConfigMetrics;
}